+++ tests/integration/tasks/test_azure_data_lake.py
@@ -1,6 +1,8 @@
 import os
 import uuid
 import pytest
+from unittest import mock
+
 
 from viadot.sources import AzureDataLake
 from viadot.tasks import (
@@ -23,6 +25,23 @@ ADLS_PATH_2 = f"raw/supermetrics/{FILE_NAME_2}"
 
 FILE_NAME_PARQUET = f"test_file_{uuid_4}.parquet"
 ADLS_PATH_PARQUET = f"raw/supermetrics/{FILE_NAME_PARQUET}"
+
+ADLS_TEST_PATHS = [
+    "raw/tests/alds_test_new_fnc/2020/02/01/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/2020/02/01/test_new_fnc.csv",
+    "raw/tests/alds_test_new_fnc/2020/02/02/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/2020/02/02/test_new_fnc.csv",
+    "raw/tests/alds_test_new_fnc/2021/12/01/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/2021/12/01/test_new_fnc.csv",
+    "raw/tests/alds_test_new_fnc/2022/06/21/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/2022/06/21/test_new_fnc.csv",
+    "raw/tests/alds_test_new_fnc/2022/08/12/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/2022/08/12/test_new_fnc.csv",
+    "raw/tests/alds_test_new_fnc/test_folder/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/test_folder/test_new_fnc.csv",
+    "raw/tests/alds_test_new_fnc/test_folder_2/final_df2.csv",
+    "raw/tests/alds_test_new_fnc/test_folder_2/test_new_fnc.csv",
+]
 
 
 @pytest.mark.dependency()
@@ -72,6 +91,23 @@ def test_azure_data_lake_list():
     list_task = AzureDataLakeList()
     files = list_task.run(path="raw/supermetrics")
     assert ADLS_PATH in files
+
+
+def test_azure_data_lake_list_recursive():
+    list_task = AzureDataLakeList()
+    files = list_task.run(path="raw/tests/alds_test_new_fnc/", recursive=True)
+    assert isinstance(files, list)
+
+
+def test_azure_data_lake_list_paths():
+
+    with mock.patch.object(
+        AzureDataLakeList, "run", return_value=ADLS_TEST_PATHS
+    ) as mock_method:
+
+        list_task = AzureDataLakeList(path="raw/tests/alds_test_new_fnc/")
+        files = list_task.run(recursive=True)
+        assert files == ADLS_TEST_PATHS
 
 
 @pytest.mark.dependency(depends=["test_azure_data_lake_upload"])