+++ viadot/sources/__init__.py
@@ -8,6 +8,8 @@ from .salesforce import Salesforce
 from .sharepoint import Sharepoint
 from .supermetrics import Supermetrics
 from .genesys import Genesys
+from .sftp import SftpConnector
+
 
 try:
     from .sap_rfc import SAPRFC
@@ -18,6 +20,7 @@ from .duckdb import DuckDB
 from .epicor import Epicor
 from .sql_server import SQLServer
 from .sqlite import SQLite
+from .mindful import Mindful
 
 # APIS
 from .uk_carbon_intensity import UKCarbonIntensity